requirements.txt
streamlit==1.31.0
boto3==1.34.25
botocore==1.34.25
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
xgboost==2.0.3
